cf2_gcs_to_bq/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud
google-cloud-bigquery